All dates with events are marked on calendar
